Split ManifoldCAD state.
