+++ bindings/wasm/examples/tsconfig.json
@@ -21,7 +21,8 @@
   },
   "include": [
     "./*.ts",
-    "./lib/*.ts"
+    "../lib/*.ts",
+    "../lib/gltf-transform/*.ts"
   ],
   "exclude": []
 }
