+++ scripts/format.sh
@@ -11,6 +11,8 @@ $CLANG_FORMAT -i test/*.{h,cpp} &
 $CLANG_FORMAT -i bindings/*/*.cpp &
 $CLANG_FORMAT -i bindings/c/include/manifold/*.h &
 $CLANG_FORMAT -i bindings/wasm/*.{js,ts} &
+$CLANG_FORMAT -i bindings/wasm/lib/*.{js,ts} &
+$CLANG_FORMAT -i bindings/wasm/lib/gltf-transform/*.{js,ts} &
 $CLANG_FORMAT -i bindings/wasm/examples/*.{js,ts,html} &
 $CLANG_FORMAT -i bindings/wasm/examples/public/*.{js,ts} &
 $CLANG_FORMAT -i src/*.{h,cpp} &
